+++ extractor.go
@@ -6,14 +6,31 @@ package gojieba
 #include "extractor.h"
 */
 import "C"
+import "unsafe"
 
 type Extractor struct {
 	extractor C.Extractor
 }
 
 func NewExtractor(dict_path, hmm_path, user_dict_path, idf_path, stop_word_path string) *Extractor {
+	dpath := C.CString(dict_path)
+	defer C.free(unsafe.Pointer(dpath))
+	hpath := C.CString(hmm_path)
+	defer C.free(unsafe.Pointer(hpath))
+	upath := C.CString(user_dict_path)
+	defer C.free(unsafe.Pointer(upath))
+	ipath := C.CString(idf_path)
+	defer C.free(unsafe.Pointer(ipath))
+	spath := C.CString(stop_word_path)
+	defer C.free(unsafe.Pointer(spath))
 	return &Extractor{
-		C.NewExtractor(C.CString(dict_path), C.CString(hmm_path), C.CString(user_dict_path), C.CString(idf_path), C.CString(stop_word_path)),
+		C.NewExtractor(
+			dpath,
+			hpath,
+			upath,
+			ipath,
+			spath,
+		),
 	}
 }
 
@@ -22,8 +39,10 @@ func (x *Extractor) Free() {
 }
 
 func (x *Extractor) Extract(s string, topk int) []string {
-	var words **C.char = C.Extract(x.extractor, C.CString(s), C.int(topk))
+	cstr := C.CString(s)
+	defer C.free(unsafe.Pointer(cstr))
+	var words **C.char = C.Extract(x.extractor, cstr, C.int(topk))
 	res := cstrings(words)
-	C.FreeWords(words)
+	defer C.FreeWords(words)
 	return res
 }