+++ ChangeLog.md
@@ -5,6 +5,7 @@
 + NewJieba(...string) support variable arguments
 + removed example/ and write Example in `*_test.go`
 + add some kind of Benchmark, Testing
++ fix bug in extractor.go: calling C.free manully to free momery allocated by C.CString
 
 ## v0.12.0